requirements.txt
Django>=3.2,<4.0
XlsxWriter>=3.0.0
django-filter>=21.0
django-pagination>=1.0.7
humanize>=4.0.0
pydicom>=2.0.0
pytz>=2021.1
celery>=5.0
python-dateutil>=2.8.0
